3yyt

+++ test.java
@@ -1,5 +1,5 @@
 public class Test {
-    private int a;
+    private float a;
     private string nom;
     public Test() {
         a = 3;
